Solved resize problem by changing to float
Prior to this commit, if you resized the swipe-pages element,
everything would go weird. This was because the width of the container
was not tracked and the display used was inline flex. The problem with
tracking the width is that it is a very costly procedure which is
unacceptable for a basic element such as this one. As such, the switch
was made to float: left; and adding an extra container in the shadow
dom. To get the widths to work correctly, the only thing that is
tracked is the number of swipe-page elements there are, which is a much
cheaper process.