Added translateZ hack and solved minor bug
1. Added translateZ hack to the swipe page container and the swipe
pages container. This guarantees (at least in webkit and blink) that
each page will be promoted to its own layer in 3d and thus will benefit
from 3d acceleration. While this sounds random, this speed up is very
noticeable on mobile when prior to this, there was a slight noticeable
lag between the position of the finger and the position of the page.
This lag has been considerable reduced and is barely noticeable now
thanks to hardware acceleration.

2. Another bug which has been solved is that of sometimes on chrome, a
page will flash white. Apparently, setting blackface-visibility to
hidden and an initial transform of scale(1) does the trick...

Basically this commit has 2 hacks that do nothing semantically to the
code and mean as much as multiplying by 1 but they dramatically improve
performance and stability... so... yay!